refactor to reduce codedup
