+++ apps/api-gateway/docker/Dockerfile.local
@@ -1,6 +1,6 @@
 # docker/dev.Dockerfile
 
-FROM oven/bun:1.2.5-slim as base
+FROM oven/bun:1.2.5-slim AS base
 
 WORKDIR /app/
 
@@ -9,15 +9,24 @@ COPY package.json ./
 COPY bun.lockb ./
 COPY tsconfig.json ./
 
-FROM base as db-build
+FROM base AS libs-build
+COPY ./packages/libs ./packages/libs
+RUN bun install
+RUN bun run libs build
+
+FROM base AS db-build
+COPY --from=libs-build /app/packages/libs/dist /app/packages/libs/dist
+COPY --from=libs-build /app/packages/libs/package.json /app/packages/libs/package.json
 COPY ./packages/db ./packages/db
 RUN bun install
 RUN bun run db build
 
 
-FROM base 
+FROM base AS runtime
 COPY --from=db-build /app/packages/db/dist /app/packages/db/dist
 COPY --from=db-build /app/packages/db/package.json /app/packages/db/package.json
+# we can copy * from libs since is already filtered
+COPY --from=db-build /app/packages/libs/ /app/packages/libs
 
 
 # Copy application code after dependencies
@@ -33,4 +42,4 @@ ENV NEXT_TELEMETRY_DISABLED=1
 RUN bun run api build
 
 # Start the application
-CMD bun run api start
+ENTRYPOINT ["bun", "run", "api", "start"]
