+++ apps/api-gateway/docker/Dockerfile.local.dockerignore
@@ -2,7 +2,8 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-/node_modules
+node_modules
+**/node_modules
 /.pnp
 .pnp.js
 
@@ -14,8 +15,10 @@
 /out/
 
 # production
-/build
-/dist
+**/build
+build
+**/dist
+dist
 
 # misc
 .DS_Store
