+++ packages/db/docker/Dockerfile.local
@@ -1,0 +1,28 @@
+# docker/dev.Dockerfile
+FROM oven/bun:1.2.5 AS base
+
+WORKDIR /app/
+
+# Copy only package files first to leverage Docker caching
+COPY package.json ./
+COPY bun.lockb ./
+COPY tsconfig.json ./
+
+
+FROM base AS libs-build
+COPY ./packages/libs ./packages/libs
+RUN bun install
+RUN bun run libs build
+# Copy application code after dependencies
+FROM base AS runtime
+COPY --from=libs-build /app/packages/libs/dist /app/packages/libs/dist
+COPY --from=libs-build /app/packages/libs/package.json /app/packages/libs/package.json
+COPY ./packages/db ./packages/db
+RUN bun install
+RUN bun run db build
+
+# Disable Next.js telemetry
+ENV NEXT_TELEMETRY_DISABLED=1
+
+# Start the application
+ENTRYPOINT ["bun", "run", "db", "typeorm", "migration:run"]
