+++ packages/db/src/env.ts
@@ -1,7 +1,7 @@
 import { Type, Static } from '@sinclair/typebox';
-import { AssertError, Value } from '@sinclair/typebox/value';
+import { loadEnv } from '@bisontrade/libs';
 
-const EnvSchema = Type.Object({
+const EnvDBSchema = Type.Object({
   NODE_ENV: Type.Union([
     Type.Literal('development'),
     Type.Literal('production'),
@@ -14,21 +14,8 @@ const EnvSchema = Type.Object({
   POSTGRES_DB: Type.String(),
 });
 
-type StaticEnv = Static<typeof EnvSchema>;
+type StaticEnv = Static<typeof EnvDBSchema>;
 
-let env: StaticEnv;
-
-try {
-  env = Value.Parse(EnvSchema, process.env);
-} catch (error) {
-  if (error instanceof AssertError) {
-    const errors = Array.from(error.Errors())
-      .filter((e) => e.type !== 45) // removing missing keys
-      .map((e) => `\t${e.path}: received ${e.value} => ${e.message}`)
-      .join('\n');
-    throw new Error(`Invalid environment:\n${errors}`);
-  }
-  throw error;
-}
+const env = loadEnv<StaticEnv>(EnvDBSchema);
 
 export default env;
