+++ packages/libs/tsconfig.json
@@ -17,6 +17,7 @@
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
 
+    "declaration": true,
     "strict": true,
     "skipLibCheck": true,
     "noFallthroughCasesInSwitch": true,
@@ -24,5 +25,6 @@
     "forceConsistentCasingInFileNames": true,
     "allowSyntheticDefaultImports": true, //do i want to keep it true??
     "allowImportingTsExtensions": false
-  }
+  },
+  "exclude": ["node_modules"]
 }
